+++ ntex-connect/src/openssl.rs
@@ -3,11 +3,11 @@ use std::{future::Future, io, pin::Pin, task::Context, task::Poll};
 pub use ntex_tls::openssl::SslFilter;
 pub use tls_openssl::ssl::{Error as SslError, HandshakeError, SslConnector, SslMethod};
 
+use ntex_bytes::PoolId;
+use ntex_io::{Base, Io};
+use ntex_service::{Service, ServiceFactory};
 use ntex_tls::openssl::SslConnector as IoSslConnector;
-
-use crate::io::{Base, Io};
-use crate::service::{Service, ServiceFactory};
-use crate::util::{PoolId, Ready};
+use ntex_util::future::Ready;
 
 use super::{Address, Connect, ConnectError, Connector as BaseConnector};
 
@@ -119,12 +119,12 @@ impl<T: Address> Service<Connect<T>> for Connector<T> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::service::{Service, ServiceFactory};
+    use ntex_service::{Service, ServiceFactory};
 
-    #[crate::rt_test]
+    #[ntex::test]
     async fn test_openssl_connect() {
-        let server = crate::server::test_server(|| {
-            crate::service::fn_service(|_| async { Ok::<_, ()>(()) })
+        let server = ntex::server::test_server(|| {
+            ntex::service::fn_service(|_| async { Ok::<_, ()>(()) })
         });
 
         let ssl = SslConnector::builder(SslMethod::tls()).unwrap();