+++ ntex-connect/src/service.rs
@@ -1,12 +1,12 @@
 use std::task::{Context, Poll};
 use std::{collections::VecDeque, future::Future, io, net::SocketAddr, pin::Pin};
 
-use crate::io::{types, Io};
-use crate::rt::tcp_connect_in;
-use crate::service::{Service, ServiceFactory};
-use crate::util::{Either, PoolId, PoolRef, Ready};
+use ntex_bytes::{PoolId, PoolRef};
+use ntex_io::{types, Io};
+use ntex_service::{Service, ServiceFactory};
+use ntex_util::future::{Either, Ready};
 
-use super::{Address, Connect, ConnectError, Resolver};
+use crate::{net::tcp_connect_in, Address, Connect, ConnectError, Resolver};
 
 pub struct Connector<T> {
     resolver: Resolver<T>,
@@ -234,10 +234,10 @@ impl<T: Address> Future for TcpConnectorResponse<T> {
 mod tests {
     use super::*;
 
-    #[crate::rt_test]
+    #[ntex::test]
     async fn test_connect() {
-        let server = crate::server::test_server(|| {
-            crate::service::fn_service(|_| async { Ok::<_, ()>(()) })
+        let server = ntex::server::test_server(|| {
+            ntex_service::fn_service(|_| async { Ok::<_, ()>(()) })
         });
 
         let srv = Connector::default().memory_pool(PoolId::P5);
@@ -256,11 +256,11 @@ mod tests {
                 .unwrap(),
             server.addr(),
         ]);
-        let result = crate::connect::connect(msg).await;
+        let result = crate::connect(msg).await;
         assert!(result.is_ok());
 
         let msg = Connect::new(server.addr());
-        let result = crate::connect::connect(msg).await;
+        let result = crate::connect(msg).await;
         assert!(result.is_ok());
     }
 }