+++ ntex-http/CHANGES.md
@@ -2,4 +2,6 @@
 
 ## [0.1.0] - 2022-xx-xx
 
+* impl FromIterator for HeaderMap
+
 * Move http types to separate crate